+++ ChangeLog.md
@@ -10,6 +10,9 @@ XP Framework Core ChangeLog
 
 ### Features
 
+* Added static `lang.Throwable::wrap()` to wrap any exception, including
+  PHP 5 and PHP 7's native base exceptions, in a `lang.Throwable` instance
+  (@thekid)
 * Changed `lang.Throwable` to accept any other instance of itself as
   cause *as well as* PHP 5 and PHP 7's native base exceptions.
   (@thekid)